Update : test_7 for number_returner Java
Changed the datatype of test_7 from 'int' to 'Double' to store values of double datatype.

+++ postfix_solver.java
@@ -436,7 +436,7 @@ public class postfix_solver
         }
 
 
-        ArrayList <int> test_7 = new ArrayList<>();
-        
+        ArrayList <Double> test_7 = new ArrayList<>();
+
     }
 }
